+++ src/main/scala/io/github/dexclaimation/overlayer/OverTransportLayer.scala
@@ -24,14 +24,14 @@ import io.github.dexclaimation.overlayer.model.Subtypes.{PID, Ref}
 import io.github.dexclaimation.overlayer.model.{PoisonPill, SchemaConfig}
 import io.github.dexclaimation.overlayer.protocol.OverWebsocket
 import io.github.dexclaimation.overlayer.protocol.common.GraphMessage._
-import io.github.dexclaimation.overlayer.protocol.common.{GraphMessage, OpMsg}
-import io.github.dexclaimation.overlayer.proxy.ProxyActions.{Connect, Disconnect, StartOp, StopOp}
-import io.github.dexclaimation.overlayer.proxy.{Proxy, ProxyActions}
+import io.github.dexclaimation.overlayer.protocol.common.{GqlError, GraphMessage, OperationMessage}
+import io.github.dexclaimation.overlayer.engine.OverActions._
+import io.github.dexclaimation.overlayer.engine.{OverActions, OverEngine}
 import sangria.execution.deferred.DeferredResolver
 import sangria.execution.{DeprecationTracker, ExceptionHandler, Middleware, QueryReducer}
 import sangria.schema.Schema
 import sangria.validation.QueryValidator
-import spray.json.{JsString, JsonParser}
+import spray.json.JsonParser
 
 import scala.concurrent.duration.{DurationInt, FiniteDuration}
 import scala.concurrent.{Await, ExecutionContext}
@@ -58,9 +58,9 @@ class OverTransportLayer[Ctx, Val](
   implicit private val ex: ExecutionContext = system.executionContext
 
 
-  private val proxy = {
-    val spawn = (rep: ActorRef[ActorRef[ProxyActions]]) => SpawnProtocol.Spawn(
-      behavior = Proxy.behavior[Ctx, Val](protocol, config),
+  private val engine = {
+    val spawn = (rep: ActorRef[ActorRef[OverActions]]) => SpawnProtocol.Spawn(
+      behavior = OverEngine.behavior[Ctx, Val](protocol, config),
       name = "ProxyStore",
       props = Props.empty,
       replyTo = rep
@@ -69,7 +69,7 @@ class OverTransportLayer[Ctx, Val](
   }
 
   private val FaultFunction: PartialFunction[String, Unit] = {
-    case PoisonPill.Pattern => new Error("Websocket connection is being shut down due to a PoisonPill Message")
+    case PoisonPill.Pattern => ()
   }
 
   /**
@@ -103,8 +103,6 @@ class OverTransportLayer[Ctx, Val](
       .toMat(Sink.asPublisher(false))(Keep.both)
       .run()
 
-    onInit(pid, actorRef)
-
     val sink: Sink[Message, Any] = Sink
       .onComplete[Unit](onEnd(pid))
       .withBefore(onMessage(ctx, pid, actorRef))
@@ -120,26 +118,37 @@ class OverTransportLayer[Ctx, Val](
 
   /** onInit Hook */
   private def onInit(pid: String, ref: Ref): InitHook = {
-    proxy ! Connect(pid, ref)
+    engine ! Connect(pid, ref)
+    protocol.init(ref)
   }
 
 
   /** onMessage Hook */
   private def onMessage(ctx: Any, pid: String, ref: Ref): MessageHook = {
     case TextMessage.Strict(msg) => JsonParser(msg).convertTo[GraphMessage] match {
-      case GraphInit() => protocol.init(ref)
+      case GraphInit() =>
+        onInit(pid, ref)
 
-      case GraphStart(oid, ast, op, vars) => proxy ! StartOp(pid, oid, ast, ctx, op, vars)
+      case GraphStart(oid, ast, op, vars) =>
+        engine ! StartOp(pid, oid, ast, ctx, op, vars)
 
-      case GraphStop(oid) => proxy ! StopOp(pid, oid)
+      case GraphImmediate(oid, ast, op, vars) =>
+        engine ! StatelessOp(pid, oid, ast, ctx, op, vars)
 
-      case GraphError(oid, message) => ref ? OpMsg.Full(protocol.error, oid, JsString(message))
+      case GraphStop(oid) =>
+        engine ! StopOp(pid, oid)
 
-      case GraphException(message) => ref ? OpMsg.NoID(protocol.error, JsString(message))
+      case GraphError(oid, message) =>
+        ref <~ OperationMessage(protocol.error, oid, GqlError.of(message))
 
-      case GraphPing() => ref ? OpMsg.Empty("pong")
+      case GraphException(message) =>
+        ref <~ OperationMessage(protocol.error, "4400", GqlError.of(message))
 
-      case GraphTerminate() => ref ! PoisonPill()
+      case GraphPing() =>
+        ref <~ OperationMessage.just("pong")
+
+      case GraphTerminate() =>
+        ref ! PoisonPill()
 
       case GraphIgnore() => ()
     }
@@ -149,7 +158,7 @@ class OverTransportLayer[Ctx, Val](
 
   /** onEnd Hook */
   private def onEnd(pid: String): EndHook = { _ =>
-    proxy ! Disconnect(pid)
+    engine ! Disconnect(pid)
   }
 
 }